VideoGLaMM/requirements.txt
packaging
sentencepiece
einops==0.4.1
fastapi==0.100.1
markdown2==2.4.10
numpy
openai==1.51.2
opencv_python
Pillow==9.4.0
pycocotools==2.0.6
ray==2.6.1
Requests==2.31.0
shortuuid==1.0.11
tqdm
uvicorn==0.23.2
scipy==1.11.2
bitsandbytes==0.41.1
timm
blobfile
mypy
pytest
requests
tensorboardX
ftfy
opencv-python
pyarrow
torchmetrics==0.7.3
pycocoevalcap
torchscale==0.2.0
hydra-core
accelerate
triton==2.1.0
peft==0.12.0
gradio==3.39.0
tensorboard
scikit-image
decord
h5py
jsonlines
ffmpeg-python
py-cpuinfo
scikit-learn
flash-attn